SW registration moved to service-worker module
