+++ core/service-worker/registration.js
